Use black and isort to reformat all files.
Also, remove a broken test case for downloading and using the tutorials.  The
current test were only causing trouble (as in, creating a 'Tutorials' directory
two levels above where the test suite is being run, or sometimes failing with
no apparent reason and a weird non-helpful message).

If we really want to have tests for the tutorials, we better start from scratch.

+++ piscat/Preproccessing/applying_mask.py
@@ -1,11 +1,11 @@
-from piscat.InputOutput.cpu_configurations import CPUConfigurations
-from joblib import Parallel, delayed
-
-import numpy as np
 import cv2 as cv
-from tqdm.autonotebook import tqdm
 import matplotlib as mpl
 import matplotlib.pylab as plt
+import numpy as np
+from joblib import Parallel, delayed
+from tqdm.autonotebook import tqdm
+
+from piscat.InputOutput.cpu_configurations import CPUConfigurations
 
 
 class Mask2Video:
@@ -24,22 +24,22 @@ class Mask2Video:
          internal_parallel_flag: bool
             Internal flag for activating parallel computation. Default is True!
         """
-        
+
         self.video = video
         self.cpu = CPUConfigurations()
         self.inter_flag_parallel_active = inter_flag_parallel_active
         if mask is not None:
             if mask.ndim == 2:
-                print('--- Same mask is used for all frame! ---')
+                print("--- Same mask is used for all frame! ---")
                 self.mask_repeater(mask)
 
             elif video.shape == mask.shape:
-                print('--- Mask has same shape with video! ---')
+                print("--- Mask has same shape with video! ---")
                 self.mask = mask
             else:
-                print('--- Mask does not have same shape with video! ---')
+                print("--- Mask does not have same shape with video! ---")
         else:
-            print('--- Mask is not define! ---')
+            print("--- Mask is not define! ---")
 
     def apply_mask(self, flag_nan=True):
         """
@@ -59,8 +59,9 @@ class Mask2Video:
         mask_video = []
         if self.inter_flag_parallel_active and self.cpu.parallel_active:
             print("\n---apply mask with Parallel---")
-            mask_video = Parallel(n_jobs=self.cpu.n_jobs, backend=self.cpu.backend, verbose=self.cpu.verbose)(delayed(
-                self.apply_mask_kernel)(f_) for f_ in tqdm(range(self.video.shape[0])))
+            mask_video = Parallel(
+                n_jobs=self.cpu.n_jobs, backend=self.cpu.backend, verbose=self.cpu.verbose
+            )(delayed(self.apply_mask_kernel)(f_) for f_ in tqdm(range(self.video.shape[0])))
             mask_video_ = np.asarray(mask_video)
 
         else:
@@ -72,8 +73,8 @@ class Mask2Video:
         if flag_nan:
             mask_video_[mask_video_ == 0] = np.nan
         else:
-            mask_video_ = mask_video_.astype('float64')
-            mask_video_[mask_video_ == 0.] = np.nan
+            mask_video_ = mask_video_.astype("float64")
+            mask_video_[mask_video_ == 0.0] = np.nan
 
             for f_ in tqdm(range(mask_video_.shape[0])):
                 img_mask = mask_video_[f_, ...]
@@ -118,5 +119,3 @@ class Mask2Video:
         cv.circle(mask, center, redius, 1, -1)
         self.mask_repeater(mask)
         return self.mask
-
-
